+++ Runtime/EventManager.cs
@@ -2,72 +2,79 @@ using System;
 
 namespace DJM.EventManager
 {
-    public class EventManager<TEventId> : EventManagerBase<TEventId>, IEventManager<TEventId>
+    public class EventManager<TEventId> : IEventManager<TEventId>
     {
+        private readonly EventService<TEventId> _eventService;
+
+        public EventManager()
+        {
+            _eventService = new EventService<TEventId>();
+        }
+
         public void AddObserver(TEventId eventId, Action handler)
         {
-            AddHandler(eventId, handler, null);
+            _eventService.AddHandler(eventId, handler, null);
         }
 
         public void AddObserver<TParam>(TEventId eventId, Action<TParam> handler)
         {
             var handlerSignature = new[] { typeof(TParam) };
-            AddHandler(eventId, handler, handlerSignature);
+            _eventService.AddHandler(eventId, handler, handlerSignature);
         }
 
         public void AddObserver<TParam1, TParam2>(TEventId eventId, Action<TParam1, TParam2> handler)
         {
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2) };
-            AddHandler(eventId, handler, handlerSignature);
+            _eventService.AddHandler(eventId, handler, handlerSignature);
         }
 
         public void AddObserver<TParam1, TParam2, TParam3>(TEventId eventId, Action<TParam1, TParam2, TParam3> handler)
         {
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2), typeof(TParam3) };
-            AddHandler(eventId, handler, handlerSignature);
+            _eventService.AddHandler(eventId, handler, handlerSignature);
         }
 
         public void AddObserver<TParam1, TParam2, TParam3, TParam4>(TEventId eventId, Action<TParam1, TParam2, TParam3, TParam4> handler)
         {
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2), typeof(TParam3), typeof(TParam4) };
-            AddHandler(eventId, handler, handlerSignature);
+            _eventService.AddHandler(eventId, handler, handlerSignature);
         }
 
         public void RemoveObserver(TEventId eventId, Action handler)
         {
-            RemoveHandler(eventId, handler, null);
+            _eventService.RemoveHandler(eventId, handler, null);
         }
 
         public void RemoveObserver<TParam>(TEventId eventId, Action<TParam> handler)
         {
             var handlerSignature = new[] { typeof(TParam) };
-            RemoveHandler(eventId, handler, handlerSignature);
+            _eventService.RemoveHandler(eventId, handler, handlerSignature);
         }
 
         public void RemoveObserver<TParam1, TParam2>(TEventId eventId, Action<TParam1, TParam2> handler)
         {
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2) };
-            RemoveHandler(eventId, handler, handlerSignature);
+            _eventService.RemoveHandler(eventId, handler, handlerSignature);
         }
 
         public void RemoveObserver<TParam1, TParam2, TParam3>(TEventId eventId, Action<TParam1, TParam2, TParam3> handler)
         {
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2), typeof(TParam3) };
-            RemoveHandler(eventId, handler, handlerSignature);
+            _eventService.RemoveHandler(eventId, handler, handlerSignature);
         }
 
         public void RemoveObserver<TParam1, TParam2, TParam3, TParam4>(TEventId eventId, Action<TParam1, TParam2, TParam3, TParam4> handler)
         {
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2), typeof(TParam3), typeof(TParam4) };
-            RemoveHandler(eventId, handler, handlerSignature);
+            _eventService.RemoveHandler(eventId, handler, handlerSignature);
         }
 
         public void TriggerEvent(TEventId eventId)
         {
-            if (!EventTable.TryGetValue(eventId, out var eventData))
+            if (!_eventService.EventTable.TryGetValue(eventId, out var eventData))
                 return;
 
-            ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, null);
+            EventService<TEventId>.ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, null);
             
             var handlers = eventData.Handlers as Action;
             handlers?.Invoke();
@@ -75,11 +82,11 @@ namespace DJM.EventManager
 
         public void TriggerEvent<TParam>(TEventId eventId, TParam param)
         {
-            if (!EventTable.TryGetValue(eventId, out var eventData))
+            if (!_eventService.EventTable.TryGetValue(eventId, out var eventData))
                 return;
             
             var handlerSignature = new[] { typeof(TParam) };
-            ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
+            EventService<TEventId>.ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
             
             var handlers = eventData.Handlers as Action<TParam>;
             handlers?.Invoke(param);
@@ -87,11 +94,11 @@ namespace DJM.EventManager
 
         public void TriggerEvent<TParam1, TParam2>(TEventId eventId, TParam1 param1, TParam2 param2)
         {
-            if (!EventTable.TryGetValue(eventId, out var eventData))
+            if (!_eventService.EventTable.TryGetValue(eventId, out var eventData))
                 return;
             
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2) };
-            ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
+            EventService<TEventId>.ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
             
             var handlers = eventData.Handlers as Action<TParam1, TParam2>;
             handlers?.Invoke(param1, param2);
@@ -99,11 +106,11 @@ namespace DJM.EventManager
 
         public void TriggerEvent<TParam1, TParam2, TParam3>(TEventId eventId, TParam1 param1, TParam2 param2, TParam3 param3)
         {
-            if (!EventTable.TryGetValue(eventId, out var eventData))
+            if (!_eventService.EventTable.TryGetValue(eventId, out var eventData))
                 return;
             
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2), typeof(TParam3) };
-            ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
+            EventService<TEventId>.ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
             
             var handlers = eventData.Handlers as Action<TParam1, TParam2, TParam3>;
             handlers?.Invoke(param1, param2, param3);
@@ -112,11 +119,11 @@ namespace DJM.EventManager
         public void TriggerEvent<TParam1, TParam2, TParam3, TParam4>(TEventId eventId, TParam1 param1, TParam2 param2, TParam3 param3,
             TParam4 param4)
         {
-            if (!EventTable.TryGetValue(eventId, out var eventData))
+            if (!_eventService.EventTable.TryGetValue(eventId, out var eventData))
                 return;
             
             var handlerSignature = new[] { typeof(TParam1), typeof(TParam2), typeof(TParam3), typeof(TParam4) };
-            ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
+            EventService<TEventId>.ValidateHandlerSignature(eventId, eventData.HandlerParamSignature, handlerSignature);
             
             var handlers = eventData.Handlers as Action<TParam1, TParam2, TParam3, TParam4>;
             handlers?.Invoke(param1, param2, param3, param4);
@@ -124,19 +131,19 @@ namespace DJM.EventManager
 
         public int GetObserverCount(TEventId eventId)
         {
-            return EventTable.TryGetValue(eventId, out var eventData)
+            return _eventService.EventTable.TryGetValue(eventId, out var eventData)
                 ? eventData.Handlers.GetInvocationList().Length
                 : 0;
         }
 
         public void ClearObservers(TEventId eventId)
         {
-            EventTable.Remove(eventId);
+            _eventService.EventTable.Remove(eventId);
         }
 
         public void ClearAllObservers()
         {
-            EventTable.Clear();
+            _eventService.EventTable.Clear();
         }
     }
 }